Bringing shell up-to-date and adding to setup.py
