+++ .travis.yml
@@ -7,10 +7,8 @@ install:
 - pipenv install --dev
 script:
 - tox
-notifications:
-- slack: iflexrtsstudio:DA5mI8cXzaRdkanMPINNpECW
 deploy:
   provider: pypi
   user: "oddjobz"
   password:
-    secure: TyZAmuwuXnN2FLOtPzKzBHqojiIcLhPO3PrAu+oNNRe98uBNw6eyZOuzPs+Nj2WzSX7GJZtJHtbb2oNbFAycIlD2sF2Od5ZIte89JBXLrGrpfgoSgNTGF82u5G+u7GL7/dSTI+RQ/UguASOuxdrpO+aUgPg3SzmgxkRhQ7sUtyEGoeJEiPUTDMMDpOqmsrjfsj0G09NLJcp2Lfz9yGWMHIfi6rL16msG9LAZGXhmwDplEqGq2zZmwEDENoX1RmXoTlBgCzsb71nXCzK6k3kkTRdeQDZX3eq8uqwn/epDJo13qEb2brL3ggZbkm67OhjcNW8of8fvmWJaywNiYpT5CdjaKTQTtDyEYu+4xAR+P3QDCZ6DTbAGzkwM5Myo5GReUDvQQK9cMPSK2kR9li06BRatVc2F8mnDe1t62ePXx/BUZpAkJCyqmqRUmWKmmGm1DYure38Lw7wgoVoDZtZrLJlpme9TSyFzkDpnjmu/FZouoQHodub5YknmTw0ffAUB3tT8MwLfbXG40hUaFvEI9Pj/k0sgNWvQx57gOyGqCHvGwo/kqGNHW23+Tmjp11uot4tj2H8g34bt/pEgulRJcCPrjQXJ7sfREwIbbt3dhuh7H9DwliXtQBfpPmexggg0EMGIvH9Dz8lFnY7Hc/elCqabfeewbGKpcxZIusn3kbo=
+    secure: LfuJ+1W9v1jIajSyh3MEOwbeXqs7q4UWUoTCcGO3FiAKhSt2TNpSTsTwB0sI25fAiRc970C/R7ftqi7ihwbhIaKmkWTIIwjJ7vyGZ3X6mhgDWafO0EwvK7CT/t6v68qQCIwZUfvgBHrgq9Cmc1jF7BYa5kxG9hcZKznHXJgUUkeUS1JNT58xalVg6yeyq9xBPq6LpwhJVWCFmUKLlpxPy4cWqhU1aY2Y/65xBN8LZ9Oj80ieyCLZA/54mmJSRCQSBwvkZJK2AnuSReVLVcHXfT02/LCNCnmsFOIuMt0a9hHuXdwEb68iBacgRh8R7S5diWi3gVJQOqk7d+IE8mhqM7otnwAKn2rs/GDKAVXc7NAerG5uhm7cxzPbFOkX4ag2sjY7po3NlzXg76QbRsNZjnJYBGyzUPOG7d/kJMuEf/osQOZIra30lUMJRMT5woAsaqgukJH399GFLVD0ithp0GhtZwG5FdUIcTTlysJAeP/iDl8zw8brW1ug5nYJEuu5p8GEyqqjs3d26KnwA1gP4fVAxIK9MLRTLGt67QskS45y9pnIzDu53JOpw86D43S7rwv5msQN4kPk3I0b4io/n2YGEaDDNPdP4LaaefXjm8RpXT59Oc1JSyRM+FIEYHz/ea1iK3xC27Yfg2AYVq5A1PiY8RJnNP/6Z++CLH87wbU=
