+++ Pipfile
@@ -1,12 +1,9 @@
 [[source]]
-
 url = "https://pypi.python.org/simple"
 verify_ssl = true
 name = "pypi"
 
-
 [packages]
-
 cmd2 = "*"
 bson = "*"
 ujson = "*"
@@ -25,10 +22,9 @@ ascii-graph = "*"
 pynndb = "*"
 cython = "*"
 pygments = "*"
-
+travis-encrypt = "*"
 
 [dev-packages]
-
 pytest = "*"
 pytest-cov = "*"
 "unittest2" = "*"
